backend/requirements.txt
# FastAPI Backend для CivilX.Universe DataLab
# Файл зависимостей Python

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# MySQL для интеграции с существующей базой пользователей
PyMySQL==1.1.0
mysql-connector-python==9.5.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0

# File Storage
minio==7.2.0
boto3==1.29.7

# Async Tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging
loguru==0.7.2

# HTTP Client
httpx==0.25.2

# Date/Time
python-dateutil==2.8.2

# IFC Processing
# ifcopenshell - версия определяется динамически из установленного пакета
# Рекомендуется установка через conda: conda install -c conda-forge ifcopenshell
# Или через pip: pip install ifcopenshell
# Или скачайте wheel файл с GitHub: https://github.com/IfcOpenShell/IfcOpenShell/releases
ifcopenshell

